@@ -155,7 +155,7 @@ func getCPUTemperatures() (map[int]float64, error) {
 	}
 
 	keys := sortFilterSysmonProperties(props, "coretemp")
-	for idx, _ := range keys {
+	for idx := range keys {
 		convertTemperatures(props[keys[idx]], res)
 	}
 
