@@ -30,11 +30,19 @@ type diskstatsCollector struct {
 	rbytes typedDesc
 	wbytes typedDesc
 	time   typedDesc
+	bsize  typedDesc
+	qdepth typedDesc
+
+	rserv typedDesc
+	wserv typedDesc
+
+	xfers typedDesc
+	xrate typedDesc
 
 	deviceFilter deviceFilter
 	logger       *slog.Logger
 
-	tickPerSecond int64
+	tickPerSecond float64
 }
 
 func init() {
@@ -57,6 +65,54 @@ func NewDiskstatsCollector(logger *slog.Logger) (Collector, error) {
 		wbytes: typedDesc{writtenBytesDesc, prometheus.CounterValue},
 		time:   typedDesc{ioTimeSecondsDesc, prometheus.CounterValue},
 
+		bsize: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "block_size_bytes"),
+				"Size of the block device in bytes.",
+				diskLabelNames, nil,
+			),
+			prometheus.GaugeValue,
+		},
+		qdepth: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "queue_depth"),
+				"Number of requests in the queue.",
+				diskLabelNames, nil,
+			),
+			prometheus.GaugeValue,
+		},
+		rserv: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "read_time_seconds_total"),
+				"The total time spent servicing read requests.",
+				diskLabelNames, nil,
+			),
+			prometheus.CounterValue,
+		},
+		wserv: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "write_time_seconds_total"),
+				"The total time spent servicing write requests.",
+				diskLabelNames, nil,
+			),
+			prometheus.CounterValue,
+		},
+		xfers: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "transfers_total"),
+				"The total number of transfers to/from disk.",
+				diskLabelNames, nil,
+			),
+			prometheus.CounterValue,
+		},
+		xrate: typedDesc{
+			prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, diskSubsystem, "transfers_to_disk_total"),
+				"The total number of transfers from disk.",
+				diskLabelNames, nil,
+			),
+			prometheus.CounterValue,
+		},
 		deviceFilter: deviceFilter,
 		logger:       logger,
 
@@ -76,7 +132,14 @@ func (c *diskstatsCollector) Update(ch chan<- prometheus.Metric) error {
 		}
 		ch <- c.rbytes.mustNewConstMetric(float64(stat.Rblks*512), stat.Name)
 		ch <- c.wbytes.mustNewConstMetric(float64(stat.Wblks*512), stat.Name)
-		ch <- c.time.mustNewConstMetric(float64(stat.Time/c.tickPerSecond), stat.Name)
+		ch <- c.time.mustNewConstMetric(float64(stat.Time)/float64(c.tickPerSecond), stat.Name)
+
+		ch <- c.bsize.mustNewConstMetric(float64(stat.BSize), stat.Name)
+		ch <- c.qdepth.mustNewConstMetric(float64(stat.QDepth), stat.Name)
+		ch <- c.rserv.mustNewConstMetric(float64(stat.Rserv)/1e9, stat.Name)
+		ch <- c.wserv.mustNewConstMetric(float64(stat.Wserv)/1e9, stat.Name)
+		ch <- c.xfers.mustNewConstMetric(float64(stat.Xfers), stat.Name)
+		ch <- c.xrate.mustNewConstMetric(float64(stat.XRate), stat.Name)
 	}
 	return nil
 }
