@@ -23,6 +23,7 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
+	"slices"
 	"strconv"
 	"strings"
 
@@ -154,11 +155,8 @@ func collectSensorData(dir string, data map[string]map[string]string) error {
 			continue
 		}
 
-		for _, t := range hwmonSensorTypes {
-			if t == sensorType {
-				addValueFile(data, sensorType+strconv.Itoa(sensorNum), sensorProperty, filepath.Join(dir, file.Name()))
-				break
-			}
+		if slices.Contains(hwmonSensorTypes, sensorType) {
+			addValueFile(data, sensorType+strconv.Itoa(sensorNum), sensorProperty, filepath.Join(dir, file.Name()))
 		}
 	}
 	return nil
