@@ -33,10 +33,10 @@ func canTestPerf(t *testing.T) {
 	if err != nil {
 		t.Skip("Procfs not mounted, skipping perf tests")
 	}
-	paranoidStr := strings.Replace(string(paranoidBytes), "\n", "", -1)
+	paranoidStr := strings.ReplaceAll(string(paranoidBytes), "\n", "")
 	paranoid, err := strconv.Atoi(paranoidStr)
 	if err != nil {
-		t.Fatalf("Expected perf_event_paranoid to be an int, got: %s", paranoidStr)
+		t.Fatalf("expected perf_event_paranoid to be an int, got: %s", paranoidStr)
 	}
 	if paranoid >= 1 {
 		t.Skip("Skipping perf tests, set perf_event_paranoid to 0")
